chore: Move fabric_automation to gen_fabric
